financial_health/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
deephaven-core==0.36.1
deephaven-plugin==0.6.0
deephaven-plugin-plotly-express==0.11.2
deephaven-plugin-utilities==0.0.2
defusedxml==0.7.1
Django==5.1.2
django-filter==24.3
djangorestframework==3.15.2
et-xmlfile==1.1.0
fastapi==0.115.2
fonttools==4.54.1
fpdf2==2.8.1
google-ai-generativelanguage==0.6.10
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.65.0
grpcio==1.67.0
grpcio-status==1.67.0
h11==0.14.0
httplib2==0.22.0
idna==3.10
jpy==0.18.0
kaleido==0.2.1
kiwisolver==1.4.7
llvmlite==0.43.0
Markdown==3.7
matplotlib==3.9.2
mysqlclient==2.2.4
numba==0.60.0
numpy==2.0.2
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
pillow==11.0.0
plotly==5.24.1
proto-plus==1.24.0
protobuf==5.28.2
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-multipart==0.0.12
pytz==2024.2
requests==2.32.3
rsa==4.9
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
starlette==0.40.0
tenacity==9.0.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
unicorn==2.1.1
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.32.0
